Refactor extensions to new config format
