+++ extension/README.md
@@ -1,0 +1,23 @@
+# Extensions
+*Note* This documentation is still in progress. For any questions, please reach
+out in the [OpenTelemetry Gitter](https://gitter.im/open-telemetry/opentelemetry-service)
+or refer to the [issues page](https://github.com/open-telemetry/opentelemetry-service/issues).
+
+Supported service extensions (sorted alphabetically):
+- [Health Check](#health-check)
+- [Performance Profiler](#pprof)
+- [zPages](#zpages)
+
+## Ordering Extensions
+The order extensions are specified for the service is important as this is the
+order in which each extension will be started and the reverse order in which they
+will be shutdown.
+
+## <a name="health-check"></a>Health Check
+TODO
+
+## <a name="pprof"></a>Performance Profiler
+TODO
+
+## <a name="zpages"></a>zPages
+TODO
