+++ extension/pprofextension/pprofextension.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pprofserver
+package pprofextension
 
 import (
 	"flag"
@@ -25,28 +25,8 @@ import (
 	"go.uber.org/zap"
 )
 
-const (
-	httpPprofPortCfg          = "http-pprof-port"
-	pprofBlockProfileFraction = "pprof-block-profile-fraction"
-	pprofMutexProfileFraction = "pprof-mutex-profile-fraction"
-)
 
-// AddFlags add the command-line flags used to control the Performance Profiler
-// (pprof) HTTP server to the given flag set.
-func AddFlags(flags *flag.FlagSet) {
-	flags.Uint(
-		httpPprofPortCfg,
-		0,
-		"Port to be used by golang net/http/pprof (Performance Profiler), the profiler is disabled if no port or 0 is specified.")
-	flags.Int(
-		pprofBlockProfileFraction,
-		-1,
-		"Fraction of blocking events that are profiled. A value <= 0 disables profiling. See runtime.SetBlockProfileRate for details.")
-	flags.Int(
-		pprofMutexProfileFraction,
-		-1,
-		"Fraction of mutex contention events that are profiled. A value <= 0 disables profiling. See runtime.SetMutexProfileFraction for details.")
-}
+
 
 // SetupFromViper sets up the Performance Profiler (pprof) as an HTTP endpoint
 // according to the configuration in the given viper.
